clang changes

+++ src/alg.cpp
@@ -5,7 +5,7 @@
 #include "alg.h"
 
 bool checkPrime(uint64_t value) {
-  for (auto i = 2; i < std::sqrt(value); ++i) {
+  for (auto i = 2; i < static_cast<int>(std::sqrt(value)); ++i) {
     if (value % i == 0) {
       return false;
     }
@@ -15,7 +15,7 @@ bool checkPrime(uint64_t value) {
 
 uint64_t nPrime(uint64_t n) {
   auto primeNumber = 2;
-  for (auto i = 1; i < n; ++i) {
+  for (auto i = 1; i < static_cast<int>(n); ++i) {
     primeNumber = nextPrime(primeNumber);
   }
   return primeNumber;
@@ -23,7 +23,7 @@ uint64_t nPrime(uint64_t n) {
 
 uint64_t nextPrime(uint64_t value) {
   auto nextPrimeNumber = 0;
-  for (auto i = value + 1; i < value * value; ++i) {
+  for (auto i = value + 1; i < static_cast<int>(value * value); ++i) {
     if (checkPrime(i)) {
       nextPrimeNumber = i;
       break;
@@ -36,7 +36,7 @@ uint64_t sumPrime(uint64_t hbound) {
   auto sum = 2;
   auto prime = 2;
   auto counter = 1;
-  while (counter != nbound) {
+  while (counter != static_cast<int>(nbound)) {
     prime = nextPrime(prime);
     sum += prime;
     counter++;
